Version que se subio al servidor

+++ areas.php
@@ -3,9 +3,17 @@ $page_title = 'Lista de áreas';
 require_once('includes/load.php');
 
 page_require_level(2);
-$all_areas = find_all_order('area','nombre_area');
+$all_areas = find_all_order('area', 'nombre_area');
 $user = current_user();
 $nivel = $user['user_level'];
+
+// $user = current_user();
+$id_usuario = $user['id'];
+
+// $user = current_user();
+$id_user = $user['id'];
+$busca_area = area_usuario($id_usuario);
+$otro = $busca_area['id'];
 //@$level = find_user_level('users', (int)$_GET['id']);
 ?>
 <?php include_once('layouts/header.php'); ?>
@@ -22,17 +30,21 @@ $nivel = $user['user_level'];
                     <span class="glyphicon glyphicon-th"></span>
                     <span>Áreas de trabajo de la CEDH</span>
                 </strong>
-                <a href="add_area.php" class="btn btn-info pull-right btn-md"> Agregar área</a>
+                <?php if ($otro == 1) : ?>
+                    <a href="add_area.php" class="btn btn-info pull-right btn-md"> Agregar área</a>
+                <?php endif ?>
             </div>
             <div class="panel-body">
-                <table class="table table-bordered table-striped">
+                <table class="datatable table table-bordered table-striped">
                     <thead>
                         <tr class="info">
                             <th class="text-center" style="width: 5%;">#</th>
                             <th style="width: 40%;">Nombre del área</th>
                             <th class="text-center" style="width: 15%;">Abreviación del área</th>
                             <th class="text-center" style="width: 20%;">Estatus del área</th>
-                            <th class="text-center" style="width: 15%;">Acciones</th>
+                            <?php if ($otro == 1) : ?>
+                                <th class="text-center" style="width: 15%;">Acciones</th>
+                            <?php endif ?>
                         </tr>
                     </thead>
                     <tbody>
@@ -48,30 +60,33 @@ $nivel = $user['user_level'];
                                         <span class="label label-danger"><?php echo "Inactiva"; ?></span>
                                     <?php endif; ?>
                                 </td>
-
-                                <td class="text-center">
-                                    <div class="btn-group">
-                                        <a href="edit_area.php?id=<?php echo (int)$a_area['id']; ?>" class="btn btn-md btn-warning" data-toggle="tooltip" title="Editar">
-                                            <i class="glyphicon glyphicon-pencil"></i>
-                                        </a>
-                                        <?php if (($nivel == 1) && ($a_area['id'] != 1)): ?>
-
-                                            <?php if ($a_area['estatus_area'] == 0)  : ?>
-                                                <a href="activate_area.php?id=<?php echo (int)$a_area['id']; ?>" class="btn btn-success btn-md" title="Activar" data-toggle="tooltip">
-                                                    <span class="glyphicon glyphicon-ok"></span>
+                                <?php if ($otro == 1) : ?>
+                                    <td class="text-center">
+                                        <div class="btn-group">
+                                            <?php if ($otro == 1) : ?>
+                                                <a href="edit_area.php?id=<?php echo (int)$a_area['id']; ?>" class="btn btn-md btn-warning" data-toggle="tooltip" title="Editar">
+                                                    <i class="glyphicon glyphicon-pencil"></i>
                                                 </a>
-                                            <?php else : ?>
-                                                <a href="inactivate_area.php?id=<?php echo (int)$a_area['id']; ?>" class="btn btn-md btn-danger" data-toggle="tooltip" title="Inactivar">
-                                                    <i class="glyphicon glyphicon-ban-circle"></i>
+                                            <?php endif ?>
+                                            <?php if (($nivel == 1) && ($a_area['id'] != 1)) : ?>
+
+                                                <?php if ($a_area['estatus_area'] == 0) : ?>
+                                                    <a href="activate_area.php?id=<?php echo (int)$a_area['id']; ?>" class="btn btn-success btn-md" title="Activar" data-toggle="tooltip">
+                                                        <span class="glyphicon glyphicon-ok"></span>
+                                                    </a>
+                                                <?php else : ?>
+                                                    <a href="inactivate_area.php?id=<?php echo (int)$a_area['id']; ?>" class="btn btn-md btn-danger" data-toggle="tooltip" title="Inactivar">
+                                                        <i class="glyphicon glyphicon-ban-circle"></i>
+                                                    </a>
+                                                <?php endif; ?>
+                                                <a href="delete_area.php?id=<?php echo (int)$a_area['id']; ?>" class="btn btn-md btn-delete" data-toggle="tooltip" title="Eliminar" onclick="return confirm('¿Seguro que deseas eliminar esta área? Los cargos de trabajo relacionados a esta se establecerán como *Sin área*.');">
+                                                    <i class="glyphicon glyphicon-trash"></i>
                                                 </a>
+
                                             <?php endif; ?>
-                                            <a href="delete_area.php?id=<?php echo (int)$a_area['id']; ?>" class="btn btn-md btn-delete" data-toggle="tooltip" title="Eliminar" onclick="return confirm('¿Seguro que deseas eliminar esta área? Los cargos de trabajo relacionados a esta se establecerán como *Sin área*.');">
-                                                <i class="glyphicon glyphicon-trash"></i>
-                                            </a>
-
-                                        <?php endif; ?>
-                                    </div>
-                                </td>
+                                        </div>
+                                    </td>
+                                <?php endif ?>
                             </tr>
                         <?php endforeach; ?>
                     </tbody>
