+++ cargos.php
@@ -7,6 +7,11 @@ $all_cargos = find_all_cargos('cargos');
 $user = current_user();
 $nivel = $user['user_level'];
 
+$id_usuario = $user['id'];
+$id_user = $user['id'];
+$busca_area = area_usuario($id_usuario);
+$otro = $busca_area['id'];
+
 ?>
 <?php include_once('layouts/header.php'); ?>
 <div class="row">
@@ -22,7 +27,9 @@ $nivel = $user['user_level'];
                     <span class="glyphicon glyphicon-th"></span>
                     <span>Cargos de la CEDH</span>
                 </strong>
-                <a href="add_cargo.php" class="btn btn-info pull-right btn-md"> Agregar cargo</a>
+                <?php if ($otro == 1) : ?>
+                    <a href="add_cargo.php" class="btn btn-info pull-right btn-md"> Agregar cargo</a>
+                <?php endif ?>
             </div>
             <div class="panel-body">
                 <table class="datatable table table-bordered table-striped">
@@ -32,7 +39,9 @@ $nivel = $user['user_level'];
                             <th style="width: 20%;">Nombre del cargo</th>
                             <th class="text-center" style="width: 30%;">Área</th>
                             <th class="text-center" style="width: 10%;">Estatus</th>
-                            <th class="text-center" style="width: 10%;">Acciones</th>
+                            <?php if ($otro == 1) : ?>
+                                <th class="text-center" style="width: 10%;">Acciones</th>
+                            <?php endif ?>
                         </tr>
                     </thead>
                     <tbody>
@@ -50,28 +59,29 @@ $nivel = $user['user_level'];
                                         <span class="label label-danger"><?php echo "Inactiva"; ?></span>
                                     <?php endif; ?>
                                 </td>
-
-                                <td class="text-center">
-                                    <div class="btn-group">
-                                        <a href="edit_cargo.php?id=<?php echo (int)$a_cargo['id']; ?>" class="btn btn-md btn-warning" data-toggle="tooltip" title="Editar">
-                                            <i class="glyphicon glyphicon-pencil"></i>
-                                        </a>
-                                        <?php if (($nivel == 1) && ($a_cargo['id'] != 1)) : ?>
-                                            <?php if ($a_cargo['estatus_cargo'] == 0) : ?>
-                                                <a href="activate_cargo.php?id=<?php echo (int)$a_cargo['id']; ?>" class="btn btn-success btn-md" title="Activar" data-toggle="tooltip">
-                                                <span class="glyphicon glyphicon-ok"></span>
-                                                </a>
-                                            <?php else : ?>
-                                                <a href="inactivate_cargo.php?id=<?php echo (int)$a_cargo['id']; ?>" class="btn btn-danger btn-md" title="Inactivar" data-toggle="tooltip">
-                                                    <span class="glyphicon glyphicon-ban-circle"></span>
+                                <?php if ($otro == 1) : ?>
+                                    <td class="text-center">
+                                        <div class="btn-group">
+                                            <a href="edit_cargo.php?id=<?php echo (int)$a_cargo['id']; ?>" class="btn btn-md btn-warning" data-toggle="tooltip" title="Editar">
+                                                <i class="glyphicon glyphicon-pencil"></i>
+                                            </a>
+                                            <?php if (($nivel == 1) && ($a_cargo['id'] != 1)) : ?>
+                                                <?php if ($a_cargo['estatus_cargo'] == 0) : ?>
+                                                    <a href="activate_cargo.php?id=<?php echo (int)$a_cargo['id']; ?>" class="btn btn-success btn-md" title="Activar" data-toggle="tooltip">
+                                                        <span class="glyphicon glyphicon-ok"></span>
+                                                    </a>
+                                                <?php else : ?>
+                                                    <a href="inactivate_cargo.php?id=<?php echo (int)$a_cargo['id']; ?>" class="btn btn-danger btn-md" title="Inactivar" data-toggle="tooltip">
+                                                        <span class="glyphicon glyphicon-ban-circle"></span>
+                                                    </a>
+                                                <?php endif; ?>
+                                                <a href="delete_cargo.php?id=<?php echo (int)$a_cargo['id']; ?>" class="btn btn-md btn-delete" data-toggle="tooltip" title="Eliminar" onclick="return confirm('¿Seguro que deseas eliminar este cargo? Los trabajadores relacionados a este cargo se establecerán como *Sin cargo*.');">
+                                                    <i class="glyphicon glyphicon-trash"></i>
                                                 </a>
                                             <?php endif; ?>
-                                            <a href="delete_cargo.php?id=<?php echo (int)$a_cargo['id']; ?>" class="btn btn-md btn-delete" data-toggle="tooltip" title="Eliminar" onclick="return confirm('¿Seguro que deseas eliminar este cargo? Los trabajadores relacionados a este cargo se establecerán como *Sin cargo*.');">
-                                                <i class="glyphicon glyphicon-trash"></i>
-                                            </a>
-                                        <?php endif; ?>
-                                    </div>
-                                </td>
+                                        </div>
+                                    </td>
+                                <?php endif ?>
                             </tr>
                         <?php endforeach; ?>
                     </tbody>
