+++ detalles_usuario.php
@@ -8,6 +8,11 @@ $all_detalles = find_all_trabajadores();
 //$all_detalles = find_all_detalles_busqueda($_POST['consulta']);
 $user = current_user();
 $nivel = $user['user_level'];
+
+$id_usuario = $user['id'];
+$id_user = $user['id'];
+$busca_area = area_usuario($id_usuario);
+$otro = $busca_area['id'];
 ?>
 <?php include_once('layouts/header.php'); ?>
 
@@ -25,7 +30,9 @@ $nivel = $user['user_level'];
           <span class="glyphicon glyphicon-th"></span>
           <span>Lista de Trabajadores de la CEDH</span>
         </strong>
-        <a href="add_detalle_usuario.php" class="btn btn-info pull-right">Agregar trabajador</a>
+        <?php if ($otro == 1) : ?>
+          <a href="add_detalle_usuario.php" class="btn btn-info pull-right">Agregar trabajador</a>
+        <?php endif ?>
       </div>
 
       <div class="panel-body">
@@ -44,7 +51,9 @@ $nivel = $user['user_level'];
               <th style="width: 5%;">Celular</th>
               <th style="width: 17%;">Área y Cargo</th>
               <th style="width: 3%;">Estatus</th>
-              <th style="width: 12%;" class="text-center">Acciones</th>
+              <?php if ($otro == 1) : ?>
+                <th style="width: 12%;" class="text-center">Acciones</th>
+              <?php endif ?>
             </tr>
           </thead>
           <tbody>
@@ -67,27 +76,29 @@ $nivel = $user['user_level'];
                     <span class="label label-danger"><?php echo "Inactivo"; ?></span>
                   <?php endif; ?>
                 </td>
-                <td class="text-center">
-                  <div class="btn-group">
-                    <a href="ver_info_detalle.php?id=<?php echo (int)$a_detalle['detalleID']; ?>" class="btn btn-md btn-info" data-toggle="tooltip" title="Ver información">
-                      <i class="glyphicon glyphicon-eye-open"></i>
-                    </a>
-                    <?php if ($nivel == 1) : ?>                      
-                      <?php if ($a_detalle['estatus_detalle'] == 0) : ?>
-                        <a href="activate_detalle_usuario.php?id=<?php echo (int)$a_detalle['detalleID']; ?>" class="btn btn-success btn-md" title="Activar" data-toggle="tooltip">
-                          <span class="glyphicon glyphicon-ok"></span>
-                        </a>
-                      <?php else : ?>
-                        <a href="inactivate_detalle_usuario.php?id=<?php echo (int)$a_detalle['detalleID']; ?>" class="btn btn-danger btn-md" title="Inactivar" data-toggle="tooltip">
-                          <span class="glyphicon glyphicon-ban-circle"></span>
+                <?php if ($otro == 1) : ?>
+                  <td class="text-center">
+                    <div class="btn-group">
+                      <a href="ver_info_detalle.php?id=<?php echo (int)$a_detalle['detalleID']; ?>" class="btn btn-md btn-info" data-toggle="tooltip" title="Ver información">
+                        <i class="glyphicon glyphicon-eye-open"></i>
+                      </a>
+                      <?php if ($nivel == 1) : ?>
+                        <?php if ($a_detalle['estatus_detalle'] == 0) : ?>
+                          <a href="activate_detalle_usuario.php?id=<?php echo (int)$a_detalle['detalleID']; ?>" class="btn btn-success btn-md" title="Activar" data-toggle="tooltip">
+                            <span class="glyphicon glyphicon-ok"></span>
+                          </a>
+                        <?php else : ?>
+                          <a href="inactivate_detalle_usuario.php?id=<?php echo (int)$a_detalle['detalleID']; ?>" class="btn btn-danger btn-md" title="Inactivar" data-toggle="tooltip">
+                            <span class="glyphicon glyphicon-ban-circle"></span>
+                          </a>
+                        <?php endif; ?>
+                        <a href="delete_detalle_usuario.php?id=<?php echo (int)$a_detalle['detalleID']; ?>" class="btn btn-delete btn-md" title="Eliminar" data-toggle="tooltip" onclick="return confirm('¿Seguro que deseas eliminar este trabajador? También se eliminarán su usuario, asignaciones y resguardos.');">
+                          <span class="glyphicon glyphicon-trash"></span>
                         </a>
                       <?php endif; ?>
-                      <a href="delete_detalle_usuario.php?id=<?php echo (int)$a_detalle['detalleID']; ?>" class="btn btn-delete btn-md" title="Eliminar" data-toggle="tooltip" onclick="return confirm('¿Seguro que deseas eliminar este trabajador? También se eliminarán su usuario, asignaciones y resguardos.');">
-                        <span class="glyphicon glyphicon-trash"></span>
-                      </a>
-                    <?php endif; ?>
-                  </div>
-                </td>
+                    </div>
+                  </td>
+                <?php endif ?>
               </tr>
             <?php endforeach; ?>
           </tbody>
